requirements.txt
python-dotenv~=0.20.0
hug~=2.6.1
requests~=2.27.1